applications/llm_ap_ar_agent/requirements.txt
langchain==0.0.331
pydantic<2.0
openai
tiktoken
faiss-cpu
python-dotenv
streamlit
chromadb
